Remove qgen package...
... as in the end, everything was inside it and nothing was
in its parent
