+++ corpus/scoll/modifiers.go
@@ -16,7 +16,7 @@
 //  You should have received a copy of the GNU General Public License
 //  along with MQUERY.  If not, see <https://www.gnu.org/licenses/>.
 
-package qgen
+package scoll
 
 import (
 	sqlLib "database/sql"
@@ -25,44 +25,43 @@ import (
 	"mquery/results"
 )
 
-// NounsModifiedByQGen
-// example:
-// Fx: [lemma="team" & deprel="nmod" & p_upos="NOUN"]
-// Fy: [p_lemma="value" & deprel="nmod" & p_upos="NOUN"]
-// Fxy: [lemma="team" & p_lemma="value" & deprel="nmod" & p_upos="NOUN"]
-type NounsModifiedByQGen struct {
+// ModifiersOfQGen
+// Fx: [p_lemma="team" & deprel="nmod" & upos="NOUN"]
+// Fy: [lemma="value" & deprel="nmod" & upos="NOUN"]
+// Fxy: [p_lemma="team" & lemma="value" & deprel="nmod" & upos="NOUN"]
+type ModifiersOfQGen struct {
 	SketchConf *CorpusSketchSetup
 }
 
-func (gen *NounsModifiedByQGen) FxQuery(word Word) string {
+func (gen *ModifiersOfQGen) FxQuery(word Word) string {
 	if word.PoS == "" {
 		return fmt.Sprintf(
 			"[%s=\"%s\" & %s=\"%s\" & %s=\"%s\"]",
-			gen.SketchConf.LemmaAttr, word.V,
+			gen.SketchConf.ParLemmaAttr, word.V,
 			gen.SketchConf.FuncAttr, gen.SketchConf.NounModifiedValue,
-			gen.SketchConf.ParPosAttr, gen.SketchConf.NounValue,
+			gen.SketchConf.PosAttr, gen.SketchConf.NounValue,
 		)
 	}
 	return fmt.Sprintf(
 		"[%s=\"%s\" & %s=\"%s\" & %s=\"%s\" & %s=\"%s\"]",
-		gen.SketchConf.LemmaAttr, word.V,
-		gen.SketchConf.PosAttr, word.PoS,
+		gen.SketchConf.ParLemmaAttr, word.V,
+		gen.SketchConf.ParPosAttr, word.PoS,
 		gen.SketchConf.FuncAttr, gen.SketchConf.NounModifiedValue,
-		gen.SketchConf.ParPosAttr, gen.SketchConf.NounValue,
+		gen.SketchConf.PosAttr, gen.SketchConf.NounValue,
 	)
 }
 
-func (gen *NounsModifiedByQGen) FxQuerySelectSQL(word Word) (sql string, args []any) {
+func (gen *ModifiersOfQGen) FxQuerySelectSQL(word Word) (sql string, args []any) {
 	if word.PoS == "" {
 		sql = fmt.Sprintf(
 			"SELECT f.result, f.result_type FROM scoll_query AS q "+
 				"JOIN scoll_fcrit AS f ON q.id = f.scoll_query_id "+
 				"WHERE q.result_type = 'Fx' AND q.%s = ? AND q.%s IS NULL AND q.%s = ? AND q.%s = ? AND f.attr = ?",
-			gen.SketchConf.LemmaAttr, gen.SketchConf.PosAttr, gen.SketchConf.FuncAttr, gen.SketchConf.ParPosAttr,
+			gen.SketchConf.ParLemmaAttr, gen.SketchConf.ParPosAttr, gen.SketchConf.FuncAttr, gen.SketchConf.PosAttr,
 		)
 		args = append(
 			args,
-			word.V, gen.SketchConf.NounModifiedValue, gen.SketchConf.NounValue, gen.SketchConf.ParLemmaAttr,
+			word.V, gen.SketchConf.NounModifiedValue, gen.SketchConf.NounValue, gen.SketchConf.LemmaAttr,
 		)
 		return
 	}
@@ -70,7 +69,7 @@ func (gen *NounsModifiedByQGen) FxQuerySelectSQL(word Word) (sql string, args []
 		"SELECT f.result, f.result_type FROM scoll_query AS q "+
 			"JOIN scoll_fcrit AS f ON q.id = f.scoll_query_id "+
 			"WHERE q.result_type = 'Fx' AND q.%s = ? AND q.%s = ? AND q.%s = ? AND q.%s = ? AND f.attr = ?",
-		gen.SketchConf.LemmaAttr, gen.SketchConf.PosAttr, gen.SketchConf.FuncAttr, gen.SketchConf.ParPosAttr,
+		gen.SketchConf.ParLemmaAttr, gen.SketchConf.ParPosAttr, gen.SketchConf.FuncAttr, gen.SketchConf.PosAttr,
 	)
 	args = append(
 		args,
@@ -79,13 +78,13 @@ func (gen *NounsModifiedByQGen) FxQuerySelectSQL(word Word) (sql string, args []
 	return
 }
 
-func (gen *NounsModifiedByQGen) FxQueryInsertSQL(word Word, result *rdb.WorkerResult) (sql string, args []any) {
+func (gen *ModifiersOfQGen) FxQueryInsertSQL(word Word, result *rdb.WorkerResult) (sql string, args []any) {
 	if result != nil && result.ResultType != results.ResultTypeFx {
-		panic(fmt.Sprintf("invalid worker result type for NounsModifiedByQGen.Fx: %s", result.ResultType))
+		panic(fmt.Sprintf("invalid worker result type for ModifiersOfQGen.Fx: %s", result.ResultType))
 	}
 	sql = fmt.Sprintf(
 		"INSERT INTO scoll_query (%s, %s, %s, %s, result, result_type) VALUES (?, ?, ?, ?, ?, ?)",
-		gen.SketchConf.LemmaAttr, gen.SketchConf.PosAttr, gen.SketchConf.FuncAttr, gen.SketchConf.ParPosAttr,
+		gen.SketchConf.ParLemmaAttr, gen.SketchConf.ParPosAttr, gen.SketchConf.FuncAttr, gen.SketchConf.PosAttr,
 	)
 	var val string
 	var rType results.ResultType
@@ -109,48 +108,48 @@ func (gen *NounsModifiedByQGen) FxQueryInsertSQL(word Word, result *rdb.WorkerRe
 	return
 }
 
-func (gen *NounsModifiedByQGen) FxCrit() string {
-	return fmt.Sprintf("%s/i 0~0>0", gen.SketchConf.ParLemmaAttr)
+func (gen *ModifiersOfQGen) FxCrit() string {
+	return fmt.Sprintf("%s/i 0~0>0", gen.SketchConf.LemmaAttr)
 }
 
-func (gen *NounsModifiedByQGen) FxCritInsertSQL(query_id int64, result *rdb.WorkerResult) (sql string, args []any) {
+func (gen *ModifiersOfQGen) FxCritInsertSQL(query_id int64, result *rdb.WorkerResult) (sql string, args []any) {
 	sql = "INSERT INTO scoll_fcrit (scoll_query_id, attr, result, result_type) VALUES (?, ?, ?, ?)"
 	args = append(
 		args,
 		query_id,
-		gen.SketchConf.ParLemmaAttr,
+		gen.SketchConf.LemmaAttr,
 		result.Value,
 		result.ResultType,
 	)
 	return
 }
 
-func (gen *NounsModifiedByQGen) FyQuery(collCandidate string) string {
+func (gen *ModifiersOfQGen) FyQuery(collCandidate string) string {
 	return fmt.Sprintf(
 		"[%s=\"%s\" & %s=\"%s\" & %s=\"%s\"]",
-		gen.SketchConf.ParLemmaAttr, collCandidate,
+		gen.SketchConf.LemmaAttr, collCandidate,
 		gen.SketchConf.FuncAttr, gen.SketchConf.NounModifiedValue,
 		gen.SketchConf.ParPosAttr, gen.SketchConf.NounValue,
 	)
 }
 
-func (gen *NounsModifiedByQGen) FyQuerySelectSQL(collCandidate string) (sql string, args []any) {
+func (gen *ModifiersOfQGen) FyQuerySelectSQL(collCandidate string) (sql string, args []any) {
 	sql = fmt.Sprintf(
 		"SELECT result, result_type FROM scoll_query "+
 			"WHERE result_type = 'Fy' AND %s = ? AND %s = ? AND %s = ?",
-		gen.SketchConf.FuncAttr, gen.SketchConf.ParPosAttr, gen.SketchConf.ParLemmaAttr,
+		gen.SketchConf.FuncAttr, gen.SketchConf.ParPosAttr, gen.SketchConf.LemmaAttr,
 	)
 	args = append(args, gen.SketchConf.NounModifiedValue, gen.SketchConf.NounValue, collCandidate)
 	return
 }
 
-func (gen *NounsModifiedByQGen) FyQueryInsertSQL(collCandidate string, result *rdb.WorkerResult) (sql string, args []any) {
+func (gen *ModifiersOfQGen) FyQueryInsertSQL(collCandidate string, result *rdb.WorkerResult) (sql string, args []any) {
 	if result.ResultType != results.ResultTypeFy {
-		panic(fmt.Sprintf("invalid worker result type for NounsModifiedByQGen.Fy: %s", result.ResultType))
+		panic(fmt.Sprintf("invalid worker result type for ModifiersOfQGen.Fy: %s", result.ResultType))
 	}
 	sql = fmt.Sprintf(
 		"INSERT INTO scoll_query (%s, %s, %s, result, result_type) VALUES (?, ?, ?, ?, ?)",
-		gen.SketchConf.ParLemmaAttr, gen.SketchConf.FuncAttr, gen.SketchConf.ParPosAttr,
+		gen.SketchConf.LemmaAttr, gen.SketchConf.FuncAttr, gen.SketchConf.ParPosAttr,
 	)
 	args = append(
 		args,
@@ -163,52 +162,59 @@ func (gen *NounsModifiedByQGen) FyQueryInsertSQL(collCandidate string, result *r
 	return
 }
 
-func (gen *NounsModifiedByQGen) FxyQuery(word Word, collCandidate string) string {
+func (gen *ModifiersOfQGen) FxyQuery(word Word, collCandidate string) string {
 	if word.PoS == "" {
 		return fmt.Sprintf(
 			"[%s=\"%s\" & %s=\"%s\" & %s=\"%s\" & %s=\"%s\"]",
-			gen.SketchConf.LemmaAttr, word.V,
-			gen.SketchConf.ParLemmaAttr, collCandidate,
+			gen.SketchConf.ParLemmaAttr, word.V,
 			gen.SketchConf.FuncAttr, gen.SketchConf.NounModifiedValue,
-			gen.SketchConf.ParPosAttr, gen.SketchConf.NounValue,
+			gen.SketchConf.PosAttr, gen.SketchConf.NounValue,
+			gen.SketchConf.LemmaAttr, collCandidate,
 		)
 	}
 	return fmt.Sprintf(
 		"[%s=\"%s\" & %s=\"%s\" & %s=\"%s\" & %s=\"%s\" & %s=\"%s\"]",
-		gen.SketchConf.LemmaAttr, word.V,
-		gen.SketchConf.PosAttr, word.PoS,
-		gen.SketchConf.ParLemmaAttr, collCandidate,
+		gen.SketchConf.ParLemmaAttr, word.V,
+		gen.SketchConf.ParPosAttr, word.PoS,
 		gen.SketchConf.FuncAttr, gen.SketchConf.NounModifiedValue,
-		gen.SketchConf.ParPosAttr, gen.SketchConf.NounValue,
+		gen.SketchConf.PosAttr, gen.SketchConf.NounValue,
+		gen.SketchConf.LemmaAttr, collCandidate,
 	)
 }
 
-func (gen *NounsModifiedByQGen) FxyQuerySelectSQL(word Word, collCandidate string) (sql string, args []any) {
+func (gen *ModifiersOfQGen) FxyQuerySelectSQL(word Word, collCandidate string) (sql string, args []any) {
 	if word.PoS == "" {
 		sql = fmt.Sprintf(
 			"SELECT result, result_type FROM scoll_query "+
 				"WHERE result_type = 'Fxy' AND %s = ? AND %s IS NULL AND %s = ? AND %s = ? AND %s = ? ",
-			gen.SketchConf.LemmaAttr, gen.SketchConf.PosAttr, gen.SketchConf.ParLemmaAttr, gen.SketchConf.FuncAttr, gen.SketchConf.ParPosAttr,
+			gen.SketchConf.ParLemmaAttr, gen.SketchConf.ParPosAttr, gen.SketchConf.FuncAttr, gen.SketchConf.ParPosAttr, gen.SketchConf.LemmaAttr,
 		)
-		args = append(args, word.V, collCandidate, gen.SketchConf.NounModifiedValue, gen.SketchConf.NounValue)
+		args = append(
+			args,
+			word.V, gen.SketchConf.NounModifiedValue, gen.SketchConf.NounValue, collCandidate,
+		)
 		return
 	}
 	sql = fmt.Sprintf(
 		"SELECT result, result_type FROM scoll_query "+
 			"WHERE result_type = 'Fxy' AND %s = ? AND %s = ? AND %s = ? AND %s = ? AND %s = ? ",
-		gen.SketchConf.LemmaAttr, gen.SketchConf.PosAttr, gen.SketchConf.ParLemmaAttr, gen.SketchConf.FuncAttr, gen.SketchConf.ParPosAttr,
+		gen.SketchConf.ParLemmaAttr, gen.SketchConf.ParPosAttr, gen.SketchConf.FuncAttr, gen.SketchConf.ParPosAttr, gen.SketchConf.LemmaAttr,
 	)
-	args = append(args, word.V, word.PoS, collCandidate, gen.SketchConf.NounModifiedValue, gen.SketchConf.NounValue)
+	args = append(
+		args,
+		word.V, word.PoS, gen.SketchConf.NounModifiedValue, gen.SketchConf.NounValue, collCandidate,
+	)
 	return
+
 }
 
-func (gen *NounsModifiedByQGen) FxyQueryInsertSQL(word Word, collCandidate string, result *rdb.WorkerResult) (sql string, args []any) {
+func (gen *ModifiersOfQGen) FxyQueryInsertSQL(word Word, collCandidate string, result *rdb.WorkerResult) (sql string, args []any) {
 	if result.ResultType != results.ResultTypeFxy {
-		panic(fmt.Sprintf("invalid worker result type for NounsModifiedByQGen.Fxy: %s", result.ResultType))
+		panic(fmt.Sprintf("invalid worker result type for ModifiersOfQGen.Fxy: %s", result.ResultType))
 	}
 	sql = fmt.Sprintf(
 		"INSERT INTO scoll_query (%s, %s, %s, %s, %s, result, result_type) VALUES (?, ?, ?, ?, ?, ?, ?)",
-		gen.SketchConf.LemmaAttr, gen.SketchConf.PosAttr, gen.SketchConf.ParLemmaAttr, gen.SketchConf.FuncAttr, gen.SketchConf.ParPosAttr,
+		gen.SketchConf.ParLemmaAttr, gen.SketchConf.ParPosAttr, gen.SketchConf.LemmaAttr, gen.SketchConf.FuncAttr, gen.SketchConf.PosAttr,
 	)
 	var posValue sqlLib.NullString
 	if word.PoS != "" {
