+++ corpus/scoll/reorder.go
@@ -16,7 +16,7 @@
 //  You should have received a copy of the GNU General Public License
 //  along with MQUERY.  If not, see <https://www.gnu.org/licenses/>.
 
-package qgen
+package scoll
 
 import (
 	"math"
