+++ corpus/scoll/verbsubj.go
@@ -16,25 +16,26 @@
 //  You should have received a copy of the GNU General Public License
 //  along with MQUERY.  If not, see <https://www.gnu.org/licenses/>.
 
-package qgen
+package scoll
 
 import (
-	sqlLib "database/sql"
 	"fmt"
 	"mquery/rdb"
 	"mquery/results"
+
+	sqlLib "database/sql"
 )
 
-type VerbObjectQGen struct {
+type VerbSubjectQGen struct {
 	SketchConf *CorpusSketchSetup
 }
 
-func (gen *VerbObjectQGen) FxQuery(word Word) string {
+func (gen *VerbSubjectQGen) FxQuery(word Word) string {
 	if word.PoS == "" {
 		return fmt.Sprintf(
 			"[%s=\"%s\" & %s=\"%s\" & %s=\"%s\"]",
 			gen.SketchConf.LemmaAttr, word.V,
-			gen.SketchConf.FuncAttr, gen.SketchConf.NounObjectValue,
+			gen.SketchConf.FuncAttr, gen.SketchConf.NounSubjectValue,
 			gen.SketchConf.ParPosAttr, gen.SketchConf.VerbValue,
 		)
 	}
@@ -42,21 +43,21 @@ func (gen *VerbObjectQGen) FxQuery(word Word) string {
 		"[%s=\"%s\" & %s=\"%s\" & %s=\"%s\" & %s=\"%s\"]",
 		gen.SketchConf.LemmaAttr, word.V,
 		gen.SketchConf.PosAttr, word.PoS,
-		gen.SketchConf.FuncAttr, gen.SketchConf.NounObjectValue,
+		gen.SketchConf.FuncAttr, gen.SketchConf.NounSubjectValue,
 		gen.SketchConf.ParPosAttr, gen.SketchConf.VerbValue,
 	)
 
 }
 
-func (gen *VerbObjectQGen) FxQuerySelectSQL(word Word) (sql string, args []any) {
+func (gen *VerbSubjectQGen) FxQuerySelectSQL(word Word) (sql string, args []any) {
 	if word.PoS == "" {
 		sql = fmt.Sprintf("SELECT f.result, f.result_type FROM scoll_query AS q "+
 			"JOIN scoll_fcrit AS f ON q.id = f.scoll_query_id "+
 			"WHERE q.result_type = 'Fx' AND q.%s = ? AND q.%s IS NULL AND q.%s = ? AND q.%s = ? AND f.attr = ?",
-			gen.SketchConf.LemmaAttr, gen.SketchConf.PosAttr, gen.SketchConf.FuncAttr, gen.SketchConf.ParPosAttr)
+			gen.SketchConf.LemmaAttr, gen.SketchConf.PosAttr, gen.SketchConf.FuncAttr, gen.SketchConf.ParLemmaAttr)
 		args = []any{
 			word.V,
-			gen.SketchConf.NounObjectValue,
+			gen.SketchConf.NounSubjectValue,
 			gen.SketchConf.VerbValue,
 			gen.SketchConf.ParLemmaAttr,
 		}
@@ -65,20 +66,20 @@ func (gen *VerbObjectQGen) FxQuerySelectSQL(word Word) (sql string, args []any) 
 	sql = fmt.Sprintf("SELECT f.result, f.result_type FROM scoll_query AS q "+
 		"JOIN scoll_fcrit AS f ON q.id = f.scoll_query_id "+
 		"WHERE q.result_type = 'Fx' AND q.%s = ? AND q.%s = ? AND q.%s = ? AND q.%s = ? AND f.attr = ?",
-		gen.SketchConf.LemmaAttr, gen.SketchConf.PosAttr, gen.SketchConf.FuncAttr, gen.SketchConf.ParPosAttr)
+		gen.SketchConf.LemmaAttr, gen.SketchConf.PosAttr, gen.SketchConf.FuncAttr, gen.SketchConf.ParLemmaAttr)
 	args = []any{
 		word.V,
 		word.PoS,
-		gen.SketchConf.NounObjectValue,
+		gen.SketchConf.NounSubjectValue,
 		gen.SketchConf.VerbValue,
 		gen.SketchConf.ParLemmaAttr,
 	}
 	return
 }
 
-func (gen *VerbObjectQGen) FxQueryInsertSQL(word Word, result *rdb.WorkerResult) (sql string, args []any) {
+func (gen *VerbSubjectQGen) FxQueryInsertSQL(word Word, result *rdb.WorkerResult) (sql string, args []any) {
 	if result != nil && result.ResultType != results.ResultTypeFx {
-		panic("invalid worker result type for VerbObjectQGen")
+		panic("invalid worker result type for VerbSubjectQGen")
 	}
 	sql = fmt.Sprintf(
 		"INSERT INTO scoll_query (%s, %s, %s, %s, result, result_type) VALUES (?, ?, ?, ?, ?, ?)",
@@ -95,10 +96,11 @@ func (gen *VerbObjectQGen) FxQueryInsertSQL(word Word, result *rdb.WorkerResult)
 		posValue.String = word.PoS
 		posValue.Valid = true
 	}
+
 	args = []any{
 		word.V,
 		posValue,
-		gen.SketchConf.NounObjectValue,
+		gen.SketchConf.NounSubjectValue,
 		gen.SketchConf.VerbValue,
 		val,
 		rType,
@@ -106,11 +108,11 @@ func (gen *VerbObjectQGen) FxQueryInsertSQL(word Word, result *rdb.WorkerResult)
 	return
 }
 
-func (gen *VerbObjectQGen) FxCrit() string {
+func (gen *VerbSubjectQGen) FxCrit() string {
 	return fmt.Sprintf("%s/i 0~0>0", gen.SketchConf.ParLemmaAttr)
 }
 
-func (gen *VerbObjectQGen) FxCritInsertSQL(query_id int64, result *rdb.WorkerResult) (sql string, args []any) {
+func (gen *VerbSubjectQGen) FxCritInsertSQL(query_id int64, result *rdb.WorkerResult) (sql string, args []any) {
 	sql = "INSERT INTO scoll_fcrit (scoll_query_id, attr, result, result_type) VALUES (?, ?, ?, ?)"
 	args = append(
 		args,
@@ -122,28 +124,28 @@ func (gen *VerbObjectQGen) FxCritInsertSQL(query_id int64, result *rdb.WorkerRes
 	return
 }
 
-func (gen *VerbObjectQGen) FyQuery(collCandidate string) string {
+func (gen *VerbSubjectQGen) FyQuery(collCandidate string) string {
 	return fmt.Sprintf(
 		`[%s="%s" & %s="%s" & %s="%s"]`,
-		gen.SketchConf.FuncAttr, gen.SketchConf.NounObjectValue,
+		gen.SketchConf.FuncAttr, gen.SketchConf.NounSubjectValue,
 		gen.SketchConf.ParPosAttr, gen.SketchConf.VerbValue,
 		gen.SketchConf.ParLemmaAttr, collCandidate,
 	)
 }
 
-func (gen *VerbObjectQGen) FyQuerySelectSQL(collCandidate string) (sql string, args []any) {
+func (gen *VerbSubjectQGen) FyQuerySelectSQL(collCandidate string) (sql string, args []any) {
 	sql = fmt.Sprintf(
 		"SELECT result, result_type FROM scoll_query "+
-			"WHERE result_type = 'Fy' AND %s = ? AND %s = ? AND %s = ?",
+			"WHERE result_type = 'Fy' AND %s = ? AND %s = ? AND %s = ? ",
 		gen.SketchConf.FuncAttr, gen.SketchConf.ParPosAttr, gen.SketchConf.ParLemmaAttr,
 	)
-	args = append(args, gen.SketchConf.NounObjectValue, gen.SketchConf.VerbValue, collCandidate)
+	args = append(args, gen.SketchConf.NounSubjectValue, gen.SketchConf.VerbValue, collCandidate)
 	return
 }
 
-func (gen *VerbObjectQGen) FyQueryInsertSQL(collCandidate string, result *rdb.WorkerResult) (sql string, args []any) {
-	if result.ResultType != results.ResultTypeFy {
-		panic("invalid worker result type for VerbObjectQGen")
+func (gen *VerbSubjectQGen) FyQueryInsertSQL(collCandidate string, result *rdb.WorkerResult) (sql string, args []any) {
+	if result != nil && result.ResultType != results.ResultTypeFy {
+		panic("invalid worker result type for VerbSubjectQGen")
 	}
 	sql = fmt.Sprintf(
 		"INSERT INTO scoll_query (%s, %s, %s, result, result_type) VALUES (?, ?, ?, ?, ?)",
@@ -151,7 +153,7 @@ func (gen *VerbObjectQGen) FyQueryInsertSQL(collCandidate string, result *rdb.Wo
 	)
 	args = append(
 		args,
-		gen.SketchConf.NounObjectValue,
+		gen.SketchConf.NounSubjectValue,
 		gen.SketchConf.VerbValue,
 		collCandidate,
 		result.Value,
@@ -160,12 +162,12 @@ func (gen *VerbObjectQGen) FyQueryInsertSQL(collCandidate string, result *rdb.Wo
 	return
 }
 
-func (gen *VerbObjectQGen) FxyQuery(word Word, collCandidate string) string {
+func (gen *VerbSubjectQGen) FxyQuery(word Word, collCandidate string) string {
 	if word.PoS == "" {
 		return fmt.Sprintf(
 			`[%s="%s" & %s="%s" & %s="%s" & %s="%s"]`,
 			gen.SketchConf.LemmaAttr, word.V,
-			gen.SketchConf.FuncAttr, gen.SketchConf.NounObjectValue,
+			gen.SketchConf.FuncAttr, gen.SketchConf.NounSubjectValue,
 			gen.SketchConf.ParPosAttr, gen.SketchConf.VerbValue,
 			gen.SketchConf.ParLemmaAttr, collCandidate,
 		)
@@ -174,38 +176,37 @@ func (gen *VerbObjectQGen) FxyQuery(word Word, collCandidate string) string {
 		`[%s="%s" & %s="%s" & %s="%s" & %s="%s" & %s="%s"]`,
 		gen.SketchConf.LemmaAttr, word.V,
 		gen.SketchConf.PosAttr, word.PoS,
-		gen.SketchConf.FuncAttr, gen.SketchConf.NounObjectValue,
+		gen.SketchConf.FuncAttr, gen.SketchConf.NounSubjectValue,
 		gen.SketchConf.ParPosAttr, gen.SketchConf.VerbValue,
 		gen.SketchConf.ParLemmaAttr, collCandidate,
 	)
 }
 
-func (gen *VerbObjectQGen) FxyQuerySelectSQL(word Word, collCandidate string) (sql string, args []any) {
+func (gen *VerbSubjectQGen) FxyQuerySelectSQL(word Word, collCandidate string) (sql string, args []any) {
 	if word.PoS == "" {
 		sql = fmt.Sprintf(
 			"SELECT result, result_type FROM scoll_query "+
-				"WHERE result_type = 'Fxy' AND %s = ? AND %s IS NULL AND %s = ? AND %s = ? AND %s = ? ",
+				" WHERE result_type = 'Fxy' AND %s = ? AND %s IS NULL AND %s = ? AND %s = ? AND %s = ? ",
 			gen.SketchConf.LemmaAttr, gen.SketchConf.PosAttr, gen.SketchConf.FuncAttr, gen.SketchConf.ParPosAttr, gen.SketchConf.ParLemmaAttr,
 		)
-		args = append(args, word.V, gen.SketchConf.NounObjectValue, gen.SketchConf.VerbValue, collCandidate)
+		args = append(args, word.V, gen.SketchConf.NounSubjectValue, gen.SketchConf.VerbValue, collCandidate)
 		return
 	}
 	sql = fmt.Sprintf(
 		"SELECT result, result_type FROM scoll_query "+
-			"WHERE result_type = 'Fxy' AND %s = ? AND %s = ? AND %s = ? AND %s = ? AND %s = ? ",
+			" WHERE result_type = 'Fxy' AND %s = ? AND %s = ? AND %s = ? AND %s = ? AND %s = ? ",
 		gen.SketchConf.LemmaAttr, gen.SketchConf.PosAttr, gen.SketchConf.FuncAttr, gen.SketchConf.ParPosAttr, gen.SketchConf.ParLemmaAttr,
 	)
 	args = append(
 		args,
-		word.V, word.PoS, gen.SketchConf.NounObjectValue, gen.SketchConf.VerbValue, collCandidate,
+		word.V, word.PoS, gen.SketchConf.NounSubjectValue, gen.SketchConf.VerbValue, collCandidate,
 	)
 	return
-
 }
 
-func (gen *VerbObjectQGen) FxyQueryInsertSQL(word Word, collCandidate string, result *rdb.WorkerResult) (sql string, args []any) {
-	if result.ResultType != results.ResultTypeFxy {
-		panic("invalid worker result type for VerbObjectQGen")
+func (gen *VerbSubjectQGen) FxyQueryInsertSQL(word Word, collCandidate string, result *rdb.WorkerResult) (sql string, args []any) {
+	if result != nil && result.ResultType != results.ResultTypeFxy {
+		panic("invalid worker result type for VerbSubjectQGen")
 	}
 	sql = fmt.Sprintf(
 		"INSERT INTO scoll_query (%s, %s, %s, %s, %s, result, result_type) VALUES (?, ?, ?, ?, ?, ?, ?)",
@@ -220,7 +221,7 @@ func (gen *VerbObjectQGen) FxyQueryInsertSQL(word Word, collCandidate string, re
 		args,
 		word.V,
 		posValue,
-		gen.SketchConf.NounObjectValue,
+		gen.SketchConf.NounSubjectValue,
 		gen.SketchConf.VerbValue,
 		collCandidate,
 		result.Value,
